Allow delegation of management of themes.
In commit 4325f9beb5 the permission of the theming control panel was
changed to allow delegation, but the browser pages used by the control
panel configlet still required the user to be a portal manager.

This commit changes the permission of these browser pages so that
delegation can work properly.